requirements.txt
starlette==0.20.4
uvicorn==0.19.0
gunicorn==20.1.0
numpy==1.23.4
fastapi==0.86.0
black==22.12.0
flake8==4.0.1
python-decouple==3.6
openai==0.25.0
elasticsearch==8.5.3